+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Changed
   each process finishes
 - Manager's communicate process only the data object by which it was
   triggered and its children.
+- Use process' slug instead of data id when logging errors in listener
 
 Added
 -----